seller inventory

+++ backend/product/models.py
@@ -12,7 +12,7 @@ class Product(models.Model):
     seller = models.ForeignKey(User, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
     quantity = models.PositiveIntegerField(null=False, default=1, blank=False)
-    price = models.FloatField(null=False, blank=False)
+    price = models.IntegerField(null=False, blank=False)
     discount_percent = models.FloatField(null=False, blank=False, default=0)
     thumbnail = models.URLField(null=True, blank=True)
